feat: extract DocsPanel component to src/components/layout/DocsPanel.tsx
Move DocsPanel, DocContent, and HealthPanel from AppView.tsx to a
dedicated file. Re-export from AppView.tsx for backward compatibility.
Remove unused imports (marked, Heart, Pause, ExternalLink, MoreHorizontal,
DropdownMenu*, Tabs*) from AppView.tsx. Add render tests covering tabs,
overflow dropdown, Health tab, and HealthPanel.
